provider: carve anthropic.rs (pure move)
The Anthropic Messages-API region moves verbatim out of
provider/mod.rs: request/response types, cache-TTL and usage
accounting, the ChatProvider impl, CU action parsing, and message
assembly, plus the 15 tests that exercise them. Moved items widen to
pub(crate); ProviderHttpResponse in mod.rs widens too (the widened
post_messages signature names it across the new boundary).
